+++ Packages/com.github.kurotu.vrc-quest-tools/Editor/Models/VRChat/AvatarConverter.cs
@@ -9,6 +9,7 @@ using System.IO;
 using System.Linq;
 using System.Runtime.ExceptionServices;
 using KRT.VRCQuestTools.Components;
+using KRT.VRCQuestTools.Importers;
 using KRT.VRCQuestTools.Models.Unity;
 using KRT.VRCQuestTools.Utils;
 using UnityEditor;
@@ -149,6 +150,11 @@ namespace KRT.VRCQuestTools.Models.VRChat
                     remover.RemoveUnsupportedComponentsInChildren(questAvatarObject, true);
                 }
                 ModularAvatarUtility.RemoveUnsupportedComponents(questAvatarObject, true);
+
+                if (saveAssetsAsFile && converterSettings.removeVertexColor)
+                {
+                    RemoveVertexColor(questAvatarObject, assetsDirectory);
+                }
 
                 ApplyVRCQuestToolsComponents(converterSettings, questAvatarObject);
 
@@ -845,21 +851,65 @@ namespace KRT.VRCQuestTools.Models.VRChat
             return convertedAnimationClips;
         }
 
+        private void RemoveVertexColor(GameObject questAvatarObject, string assetsDirectory)
+        {
+            var renderers = questAvatarObject.GetComponentsInChildren<SkinnedMeshRenderer>(true)
+                .Cast<Renderer>()
+                .Concat(questAvatarObject.GetComponentsInChildren<MeshRenderer>(true))
+                .Where(r =>
+                {
+                    var mesh = RendererUtility.GetSharedMesh(r);
+                    return mesh != null && mesh.colors32 != null && mesh.colors32.Length > 0;
+                });
+
+            if (!renderers.Any())
+            {
+                return;
+            }
+
+            var saveDirectory = $"{assetsDirectory}/Meshes";
+            Directory.CreateDirectory(saveDirectory);
+
+            var convertedMeshes = new Dictionary<Mesh, Mesh>();
+
+            foreach (var renderer in renderers)
+            {
+                var mesh = RendererUtility.GetSharedMesh(renderer);
+
+                if (convertedMeshes.ContainsKey(mesh))
+                {
+                    RendererUtility.SetSharedMesh(renderer, convertedMeshes[mesh]);
+                    continue;
+                }
+
+                var originalName = mesh.name;
+                AssetDatabase.TryGetGUIDAndLocalFileIdentifier(mesh, out string guid, out long _);
+                var outFile = $"{saveDirectory}/{originalName}_from_{guid}.vqtmesh";
+
+                // When the mesh is added into another asset, "/" is acceptable as name.
+                if (originalName.Contains("/"))
+                {
+                    var dir = Path.GetDirectoryName(outFile);
+                    Directory.CreateDirectory(dir);
+                }
+
+                MeshModifierImporter.CreateAsset(mesh, outFile, true);
+                var newMesh = AssetDatabase.LoadAssetAtPath<Mesh>(outFile);
+                if (newMesh == null)
+                {
+                    Logger.LogError($"Failed to load generated mesh asset at path '{outFile}' for renderer '{renderer.name}'.");
+                    continue;
+                }
+                convertedMeshes[mesh] = newMesh;
+                RendererUtility.SetSharedMesh(renderer, newMesh);
+            }
+        }
+
         private void ApplyVRCQuestToolsComponents(AvatarConverterSettings setting, GameObject questAvatarObject)
         {
             if (questAvatarObject.GetComponent<ConvertedAvatar>() == null)
             {
                 questAvatarObject.AddComponent<ConvertedAvatar>();
-            }
-            if (setting.removeVertexColor)
-            {
-                var vcr = questAvatarObject.GetComponent<VertexColorRemover>();
-                if (vcr == null)
-                {
-                    vcr = questAvatarObject.AddComponent<VertexColorRemover>();
-                }
-                vcr.includeChildren = true;
-                vcr.enabled = true;
             }
 
 #if VQT_HAS_NDMF